+++ packages/editor/web/src/RichContentEditor/Decorators/Resize/index.js
@@ -1,15 +1,4 @@
-import { EditorState } from 'draft-js';
 import createDecorator from './createDecorator';
-
-const createSetResizeData = (contentBlock, { getEditorState, setEditorState }) => data => {
-  const entityKey = contentBlock.getEntityAt(0);
-  if (entityKey) {
-    const editorState = getEditorState();
-    const contentState = editorState.getCurrentContent();
-    contentState.mergeEntityData(entityKey, { ...data });
-    setEditorState(EditorState.forceSelection(editorState, editorState.getSelection()));
-  }
-};
 
 export default config => {
   const store = {
@@ -26,16 +15,5 @@ export default config => {
       store.setEditorState = setEditorState;
     },
     decorator: createDecorator({ config, store }),
-    blockRendererFn: (contentBlock, { getEditorState, setEditorState }) => {
-      const entityKey = contentBlock.getEntityAt(0);
-      const contentState = getEditorState().getCurrentContent();
-      const resizeData = entityKey ? contentState.getEntity(entityKey).data : {};
-      return {
-        props: {
-          resizeData,
-          setResizeData: createSetResizeData(contentBlock, { getEditorState, setEditorState }),
-        },
-      };
-    },
   };
 };