+++ CHANGELOG.md
@@ -38,9 +38,10 @@
     - [#446](https://github.com/wix-incubator/rich-content/pull/446) `pro-gallery` updated to `v1.5.25`
   - `e2e`
     - [#447](https://github.com/wix-incubator/rich-content/pull/447) added more supported browsers
+  - `resize`
+    - [#450](https://github.com/wix-incubator/rich-content/pull/450) moved width from data into data.config
   - `babel`
-    - [#454](https://github.com/wix-incubator/rich-content/pull/454) add optional chaining support
-    
+    - [#454](https://github.com/wix-incubator/rich-content/pull/454) add optional chaining support   
 </details>
 <hr/>